requirements.txt
google-cloud-storage
uvicorn
fastapi
simpletransformers
torch
